refactor: split product queries and repository read modules
Close the broad domain read-model and live-archive stewardship campaign by decomposing model roots, repository reads, and session-product queries; adding archive-debt governance and public analytics/debt products; and wiring live validation lanes and public consumer surfaces.\n\n🤖 Generated with [Claude Code](https://claude.com/claude-code)\n\nCo-Authored-By: Claude <noreply@anthropic.com>
