+++ devtools/run_validation_lanes.py
@@ -294,6 +294,38 @@ LANES: dict[str, LaneConfig] = {
             "--json",
         ],
     ),
+    "live-products-analytics": LaneConfig(
+        name="live-products-analytics",
+        description="Live archive provider-analytics product surface",
+        timeout_s=180,
+        command=[
+            sys.executable,
+            "-m",
+            "polylogue",
+            "--plain",
+            "products",
+            "analytics",
+            "--limit",
+            "20",
+            "--json",
+        ],
+    ),
+    "live-products-debt": LaneConfig(
+        name="live-products-debt",
+        description="Live archive debt/governance product surface",
+        timeout_s=180,
+        command=[
+            sys.executable,
+            "-m",
+            "polylogue",
+            "--plain",
+            "products",
+            "debt",
+            "--limit",
+            "20",
+            "--json",
+        ],
+    ),
     "live-project-stats": LaneConfig(
         name="live-project-stats",
         description="Live archive project-grouped stats over durable session products",
@@ -434,6 +466,29 @@ LANES: dict[str, LaneConfig] = {
         description="Local durable-product contract lane plus bounded live archive product dogfooding",
         timeout_s=1800,
         sub_lanes=("archive-data-products", "live-products-small"),
+    ),
+    "domain-read-model-contracts": LaneConfig(
+        name="domain-read-model-contracts",
+        description="Local domain-read-model closure lane for analytics/products, consumer contracts, and archive-debt governance surfaces",
+        timeout_s=2400,
+        sub_lanes=("archive-data-products", "maintenance-control-plane"),
+    ),
+    "domain-read-model-live": LaneConfig(
+        name="domain-read-model-live",
+        description="Bounded live archive lane for durable products, analytics/debt products, and maintenance governance",
+        timeout_s=1800,
+        sub_lanes=(
+            "live-products-small",
+            "live-products-analytics",
+            "live-products-debt",
+            "live-governance-small",
+        ),
+    ),
+    "domain-read-model-stewardship": LaneConfig(
+        name="domain-read-model-stewardship",
+        description="Full domain-read-model and live-archive stewardship closure lane",
+        timeout_s=3600,
+        sub_lanes=("domain-read-model-contracts", "domain-read-model-live"),
     ),
     "runtime-substrate-contracts": LaneConfig(
         name="runtime-substrate-contracts",
