+++ docs/cli-reference.md
@@ -216,6 +216,8 @@ polylogue products threads --limit 20 --json
 polylogue products tags --provider claude-code --since 2026-01-01 --json
 polylogue products day-summaries --provider claude-code --since 2026-01-01 --json
 polylogue products week-summaries --provider claude-code --since 2026-01-01 --json
+polylogue products analytics --provider claude-code --json
+polylogue products debt --actionable-only --json
 polylogue products maintenance --json
 ```
 
@@ -223,9 +225,12 @@ Durable archive products expose versioned, machine-readable session/work and mai
 
 The session-product surfaces now include:
 
+- `products status`: durable session-product readiness plus archive-debt summary
 - `products profiles`: `first_message_at`, `canonical_session_date`, `engaged_duration_ms`, `engaged_minutes`, canonical projects, repo paths
 - `products work-events`: timestamped work-event rows with `start_time`, `end_time`, `duration_ms`, `canonical_session_date`
 - `products phases`: timestamped session-phase timeline rows with phase kind, timing, duration, and per-phase tool counts
+- `products analytics`: stable provider-level conversation/message/tool/thinking metrics
+- `products debt`: governed archive-cleanup and derived-repair debt items with maintenance targets
 
 ## Tags
