+++ polylogue/operations/__init__.py
@@ -1,17 +1,17 @@
 """Canonical archive operations shared across facade, CLI, and MCP surfaces."""
 
 from .archive import (
+    ArchiveDebtProduct,
     ArchiveOperations,
     ArchiveStats,
-    ProviderMetrics,
-    compute_provider_comparison,
     get_provider_counts,
+    list_provider_analytics_products,
 )
 
 __all__ = [
+    "ArchiveDebtProduct",
     "ArchiveOperations",
     "ArchiveStats",
-    "ProviderMetrics",
-    "compute_provider_comparison",
     "get_provider_counts",
+    "list_provider_analytics_products",
 ]
