+++ polylogue/storage/repository.py
@@ -20,14 +20,20 @@ if TYPE_CHECKING:
 
     from polylogue.storage.backends.query_store import SQLiteQueryStore
 
+from polylogue.storage.repository_action_reads import RepositoryActionReadMixin
+from polylogue.storage.repository_archive_reads import RepositoryArchiveReadMixin
+from polylogue.storage.repository_maintenance_reads import RepositoryMaintenanceReadMixin
+from polylogue.storage.repository_product_reads import RepositoryProductReadMixin
 from polylogue.storage.repository_raw import RepositoryRawMixin
-from polylogue.storage.repository_reads import RepositoryReadMixin
 from polylogue.storage.repository_vectors import RepositoryVectorMixin
 from polylogue.storage.repository_writes import RepositoryWriteMixin
 
 
 class ConversationRepository(
-    RepositoryReadMixin,
+    RepositoryArchiveReadMixin,
+    RepositoryActionReadMixin,
+    RepositoryMaintenanceReadMixin,
+    RepositoryProductReadMixin,
     RepositoryRawMixin,
     RepositoryWriteMixin,
     RepositoryVectorMixin,
