+++ polylogue/storage/repository_archive_reads.py
@@ -1,4 +1,4 @@
-"""Read/query method mixin for the conversation repository."""
+"""Archive conversation/message/tree read methods for the repository."""
 
 from __future__ import annotations
 
@@ -7,29 +7,23 @@ import builtins
 from collections.abc import AsyncIterator
 from typing import TYPE_CHECKING
 
-from polylogue.lib.models import Conversation, ConversationSummary, Message
-from polylogue.storage.action_event_rows import hydrate_action_events
+from polylogue.lib.conversation_models import Conversation, ConversationSummary
+from polylogue.lib.message_models import Message
 from polylogue.storage.hydrators import (
     conversation_from_records,
     conversation_summary_from_record,
     message_from_record,
 )
 from polylogue.storage.query_models import ConversationRecordQuery
-from polylogue.storage.session_product_rows import (
-    hydrate_session_phase,
-    hydrate_session_profile,
-    hydrate_work_event,
-    hydrate_work_thread,
-)
 from polylogue.storage.state_views import ConversationRenderProjection
-from polylogue.storage.store import ActionEventRecord, ConversationRecord
+from polylogue.storage.store import ConversationRecord
 from polylogue.types import ConversationId
 
 if TYPE_CHECKING:
     from polylogue.lib import filters
 
 
-class RepositoryReadMixin:
+class RepositoryArchiveReadMixin:
     async def resolve_id(self, id_prefix: str) -> ConversationId | None:
         resolved = await self.queries.resolve_id(id_prefix)
         return ConversationId(resolved) if resolved else None
@@ -43,7 +37,6 @@ class RepositoryReadMixin:
             self.queries.get_messages(conversation_id),
             self.queries.get_attachments(conversation_id),
         )
-
         return conversation_from_records(conv_record, msg_records, att_records)
 
     async def get_render_projection(self, conversation_id: str) -> ConversationRenderProjection | None:
@@ -107,7 +100,7 @@ class RepositoryReadMixin:
         query: ConversationRecordQuery,
     ) -> builtins.list[ConversationSummary]:
         conv_records = await self.queries.list_conversations(query)
-        return [conversation_summary_from_record(rec) for rec in conv_records]
+        return [conversation_summary_from_record(record) for record in conv_records]
 
     async def list_summaries(
         self,
@@ -354,7 +347,7 @@ class RepositoryReadMixin:
         ids, records = await self._search_records(query, limit=limit, providers=providers)
         if not ids:
             return []
-        return [conversation_summary_from_record(rec) for rec in records]
+        return [conversation_summary_from_record(record) for record in records]
 
     async def search(
         self,
@@ -405,290 +398,8 @@ class RepositoryReadMixin:
     async def get_many(self, conversation_ids: builtins.list[str]) -> builtins.list[Conversation]:
         if not conversation_ids:
             return []
-
         records = await self.queries.get_conversations_batch(conversation_ids)
         return await self._hydrate_conversations(records, ordered_ids=conversation_ids)
-
-    async def get_action_event_records(self, conversation_id: str) -> list[ActionEventRecord]:
-        return await self.queries.get_action_events(conversation_id)
-
-    async def get_action_event_records_batch(
-        self,
-        conversation_ids: builtins.list[str],
-    ) -> dict[str, list[ActionEventRecord]]:
-        return await self.queries.get_action_events_batch(conversation_ids)
-
-    async def get_action_events(self, conversation_id: str):
-        return hydrate_action_events(await self.get_action_event_records(conversation_id))
-
-    async def get_action_events_batch(
-        self,
-        conversation_ids: builtins.list[str],
-    ) -> dict[str, tuple]:
-        records_by_conversation = await self.get_action_event_records_batch(conversation_ids)
-        return {
-            conversation_id: hydrate_action_events(records)
-            for conversation_id, records in records_by_conversation.items()
-        }
-
-    async def get_action_event_read_model_status(self) -> dict[str, int | bool]:
-        return await self.queries.get_action_event_read_model_status()
-
-    async def get_session_product_status(self) -> dict[str, int | bool]:
-        return await self.queries.get_session_product_status()
-
-    async def get_session_profile_record(self, conversation_id: str):
-        return await self.queries.get_session_profile(conversation_id)
-
-    async def get_session_profile(self, conversation_id: str):
-        record = await self.get_session_profile_record(conversation_id)
-        return hydrate_session_profile(record) if record is not None else None
-
-    async def get_session_profile_records_batch(
-        self,
-        conversation_ids: builtins.list[str],
-    ):
-        return await self.queries.get_session_profiles_batch(conversation_ids)
-
-    async def get_session_profiles_batch(
-        self,
-        conversation_ids: builtins.list[str],
-    ):
-        records = await self.get_session_profile_records_batch(conversation_ids)
-        return {
-            conversation_id: hydrate_session_profile(record)
-            for conversation_id, record in records.items()
-        }
-
-    async def list_session_profiles(
-        self,
-        *,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        first_message_since: str | None = None,
-        first_message_until: str | None = None,
-        session_date_since: str | None = None,
-        session_date_until: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        records = await self.queries.list_session_profiles(
-            provider=provider,
-            since=since,
-            until=until,
-            first_message_since=first_message_since,
-            first_message_until=first_message_until,
-            session_date_since=session_date_since,
-            session_date_until=session_date_until,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-        return [hydrate_session_profile(record) for record in records]
-
-    async def list_session_profile_records(
-        self,
-        *,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        first_message_since: str | None = None,
-        first_message_until: str | None = None,
-        session_date_since: str | None = None,
-        session_date_until: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        return await self.queries.list_session_profiles(
-            provider=provider,
-            since=since,
-            until=until,
-            first_message_since=first_message_since,
-            first_message_until=first_message_until,
-            session_date_since=session_date_since,
-            session_date_until=session_date_until,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-
-    async def get_session_work_event_records(self, conversation_id: str):
-        return await self.queries.get_session_work_events(conversation_id)
-
-    async def get_session_phase_records(self, conversation_id: str):
-        return await self.queries.get_session_phases(conversation_id)
-
-    async def get_session_work_events(self, conversation_id: str):
-        return [hydrate_work_event(record) for record in await self.get_session_work_event_records(conversation_id)]
-
-    async def get_session_phases(self, conversation_id: str):
-        return [hydrate_session_phase(record) for record in await self.get_session_phase_records(conversation_id)]
-
-    async def list_session_work_events(
-        self,
-        *,
-        conversation_id: str | None = None,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        kind: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        records = await self.queries.list_session_work_events(
-            conversation_id=conversation_id,
-            provider=provider,
-            since=since,
-            until=until,
-            kind=kind,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-        return [hydrate_work_event(record) for record in records]
-
-    async def list_session_work_event_records(
-        self,
-        *,
-        conversation_id: str | None = None,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        kind: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        return await self.queries.list_session_work_events(
-            conversation_id=conversation_id,
-            provider=provider,
-            since=since,
-            until=until,
-            kind=kind,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-
-    async def list_session_phases(
-        self,
-        *,
-        conversation_id: str | None = None,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        kind: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-    ):
-        records = await self.queries.list_session_phases(
-            conversation_id=conversation_id,
-            provider=provider,
-            since=since,
-            until=until,
-            kind=kind,
-            limit=limit,
-            offset=offset,
-        )
-        return [hydrate_session_phase(record) for record in records]
-
-    async def list_session_phase_records(
-        self,
-        *,
-        conversation_id: str | None = None,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        kind: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-    ):
-        return await self.queries.list_session_phases(
-            conversation_id=conversation_id,
-            provider=provider,
-            since=since,
-            until=until,
-            kind=kind,
-            limit=limit,
-            offset=offset,
-        )
-
-    async def get_work_thread_record(self, thread_id: str):
-        return await self.queries.get_work_thread(thread_id)
-
-    async def get_work_thread(self, thread_id: str):
-        record = await self.get_work_thread_record(thread_id)
-        return hydrate_work_thread(record) if record is not None else None
-
-    async def list_work_threads(
-        self,
-        *,
-        since: str | None = None,
-        until: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        records = await self.queries.list_work_threads(
-            since=since,
-            until=until,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-        return [hydrate_work_thread(record) for record in records]
-
-    async def list_work_thread_records(
-        self,
-        *,
-        since: str | None = None,
-        until: str | None = None,
-        limit: int | None = 50,
-        offset: int = 0,
-        query: str | None = None,
-    ):
-        return await self.queries.list_work_threads(
-            since=since,
-            until=until,
-            limit=limit,
-            offset=offset,
-            query=query,
-        )
-
-    async def list_session_tag_rollup_records(
-        self,
-        *,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-        query: str | None = None,
-    ):
-        return await self.queries.list_session_tag_rollup_rows(
-            provider=provider,
-            since=since,
-            until=until,
-            query=query,
-        )
-
-    async def list_day_session_summary_records(
-        self,
-        *,
-        provider: str | None = None,
-        since: str | None = None,
-        until: str | None = None,
-    ):
-        return await self.queries.list_day_session_summaries(
-            provider=provider,
-            since=since,
-            until=until,
-        )
-
-    async def list_maintenance_runs(self, *, limit: int = 20):
-        return await self.queries.list_maintenance_runs(limit=limit)
 
     async def iter_messages(
         self,
@@ -710,3 +421,6 @@ class RepositoryReadMixin:
         from polylogue.lib import filters
 
         return filters.ConversationFilter(self)
+
+
+__all__ = ["RepositoryArchiveReadMixin"]
